global-schemas/requirements.txt
arize_phoenix==3.24.0
chromadb==0.5.0
evaluate==0.4.1
hdbscan
ipykernel
Flask-SQLAlchemy==3.1.1
groq==0.5.0
langchain==0.1.16
langchain_groq==0.1.3
langchain-openai==0.1.4
langid==1.1.6
lightgbm==4.3.0
nltk==3.8.1
numpy==1.26.4
omegaconf==2.3.0
openai==1.23.6
pandas==2.2.2
psycopg2-binary==2.9.9
peft==0.10.0
pydantic==2.8.2
python-dotenv==1.0.1
scikit-learn==1.4.2
selenium==4.20.0
chromedriver-py==126.0.6478.182
seaborn==0.13.2
sentence_transformers==2.7.0
shap==0.45.0
stem==1.8.2
SQLAlchemy==2.0.29
sqlparse==0.5.0
swifter==1.4.0
timm==0.9.16
tqdm==4.66.2
trl==0.8.6
umap-learn==0.5.6
unidecode==1.3.8
yfinance==0.2.38
